% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate data}
\usage{
aggregate_results(
  project = project,
  quantile = NA,
  file_name = NA,
  reread = FALSE,
  save_project_steps = FALSE
)
}
\arguments{
\item{project}{Name of the initial project in the \code{environment}.
The default value is \code{project}.}

\item{quantile}{Value of the lowest quantile to aggregate the data.
Default value: \code{NA}. For example: \code{0.05}}

\item{file_name}{Custom file name (without extension) for the exported
results. Default value: \code{NA}.}

\item{reread}{Read the longer \code{csv} table from \code{project_folder}
and process it. For example to exclude compounds or single results.
Default value: \code{FALSE} (values: \code{c(FALSE, TRUE)}).}

\item{save_project_steps}{Stores the current project state in the \code{project_folder}.
The default value is \code{FALSE} (values: \code{c(FALSE, TRUE)}).}
}
\description{
This function aggregates the longer formmat results of the EcoTox
Knowledgebase query stored in the \code{REcoTox} project to the
final longer results data table.
}
\examples{
# Run the current project with a quantile of {0.05} with a custom name.

\dontrun{aggregate_results(project = project, quantile = 0.05,
file_name = "algae_aggregated_results")}

}
\author{
Tobias Schulze
}
