% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create project}
\usage{
create_project(
  database_path,
  project_path,
  initalise_database_project = FALSE,
  initalise_project = FALSE,
  load_default = FALSE
)
}
\arguments{
\item{database_path}{Path to the folder containing the expanded ascii files
of the current \code{EcoTox Knowledgebase}.}

\item{project_path}{Path to the folder to store the processed project files.}

\item{initalise_database_project}{Logical value to initialize a new default
\code{REcoTox} project in the \code{database_path}. The default value is
\code{FALSE} (values: \code{c(TRUE, FALSE)}).}

\item{initalise_project}{Logical value zu initialize the project folder in
the \code{project_path}. The default value is \code{FALSE}
(values:\code{c(TRUE, FALSE)}).}

\item{load_default}{Logical value to load the default project in the
\code{database_path}. The default value is \code{FALSE} (values:
\code{c(TRUE, FALSE)}. The parameter is ignored, if the \code{initialise_
project} parameter is \code{TRUE}.}
}
\description{
This function initializes the \code{REcoTox} project in
the pre-defined \code{project_folder} in \code{project_path}.
It reads also the \code{EcoTox Knowledgebase} \emph{ascii}-files
provided in \code{database_path} and prepares the
\code{REcoTox} project.

The prepared project is exported to \code{database_path} as default
project for re-use in other projects based on the same \code{EcoTox}
database version.
}
\examples{
# Initialize a new default \code{REcoTox} project in
# the \code{database_path} by reading the \code{
# EcoTox Knowledgebase} \emph{ascii}-files.

\dontrun{
project <- create_project(database_path = "/extdata/database_folder",
project_path = "/extdata/project_folder",
initalise_database_project = TRUE
)}

}
\author{
Tobias Schulze
}
